laporan kurang export pdf

+++ database/migrations/2022_08_16_045406_create_brilian_detail_table.php
@@ -20,7 +20,7 @@ class CreateBrilianDetailTable extends Migration
             $table->string('kode_mk');
             $table->string('kelas');
             $table->string('prodi');
-            $table->integer('nilai');
+            $table->decimal('nilai',6,2);
 
             $table->foreign('brilian_week_id')->references('id')->on('brilian_week')->onDelete('cascade');
             $table->timestamps();
